DRA7xx: power on DSP when dsptop starts
DRA7xx (AM57x) might have power control that automatically suspends
and resumes DSPs to save power consumption.  However, if DSP is suspended,
dsptop won't be able to access ETB on the DSP side.  The fix is to switch
power control on DSP to "always on" when dsptop starts, and to its
original status (either "on" or "auto") when dsptop exits.